Merged PR 203: refactor: move attribute state and behaviour to individual vuex-modules
Create a new vuex module for each attribute. These new modules will contain the state (selected, loading...) as well as the data fetching logic via Actions.

The reason for the refactoring is to better separate pieces of functionality, thereby making it easier to work with with more complex requirements.

Related work items: #768